try radix sort optimization

+++ push_swap/src/main.c
@@ -6,7 +6,7 @@
 /*   By: rmander <rmander@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/30 23:08:41 by rmander           #+#    #+#             */
-/*   Updated: 2021/08/25 22:22:30 by rmander          ###   ########.fr       */
+/*   Updated: 2021/08/26 00:16:24 by rmander          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -103,9 +103,9 @@ void	printop(void *op)
 	write(STDOUT_FILENO, &endl, 1);
 }
 
-int *indexify(t_data *data, int *values, int size)
+int *indexify(t_data *data, int *values, size_t size)
 {
-	int		i;
+	size_t	i;
 	int		*copy;
 
 	i = 0;
@@ -185,6 +185,7 @@ size_t	ft_lstsize(t_list *lst)
 	return (size);
 }
 
+
 int main(int argc, char **argv)
 {
 	t_data	data;
@@ -208,13 +209,12 @@ int main(int argc, char **argv)
 	data.b = build(NULL, argc);
 	if (!data.b)
 		pexit(&data, EXIT_FAILURE);
+
 	if (!empty(data.a))
 		push_swap(&data);
 
 	/* operations optimizations */
-
 	size_t	nops;
-
 	nops = ft_lstsize(data.ops);
 	while (nops--)
 		optimize(&data);
